Moved AwsLambdaRequestEvent to events folder
