+++ aws-lambda-java-events/src/main/java/com/amazonaws/services/lambda/runtime/events/AwsLambdaRequestEvent.java
@@ -1,4 +1,6 @@
-package com.amazonaws.services.lambda.runtime.events.helper;
+package com.amazonaws.services.lambda.runtime.events;
+
+import com.amazonaws.services.lambda.runtime.events.helper.RequestSource;
 
 /**
  * Interface to abstract APIGWY and ALB request classes, reduce complexity and increase efficiency.
