+++ aws-lambda-java-events/src/main/java/com/amazonaws/services/lambda/runtime/events/apigateway/APIGatewayV2HTTPEvent.java
@@ -13,7 +13,7 @@
 
 package com.amazonaws.services.lambda.runtime.events.apigateway;
 
-import com.amazonaws.services.lambda.runtime.events.helper.AwsLambdaRequestEvent;
+import com.amazonaws.services.lambda.runtime.events.AwsLambdaRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.helper.RequestSource;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
